a lot of debbuging and errors handling need to end strings table and start doing some useful functions

+++ db.py
@@ -1,5 +1,5 @@
 import psycopg2
-
+from datetime import datetime
 conn = psycopg2.connect(dbname="postgres", host="localhost", user="postgres", password="12345", port="5432")
 cursor = conn.cursor()
 
@@ -27,6 +27,7 @@ class educationsTable():
         except (psycopg2.errors.ForeignKeyViolation):
             conn.commit()
             return "FK error"
+        
     def update_education_table(edu_id, eduType):
         if edu_id not in [i[0] for i in educationsTable.show_educations_table()]:
             conn.commit()
@@ -111,16 +112,23 @@ class employeesTable():
     def add_to_employees_table(surname, phone_num, edu_id):
         try:
             cursor.execute("INSERT INTO employees (surname, phone_number, edu_id) VALUES (%s,%s,%s)", (surname,phone_num, edu_id))
-        except Exception as e:
-            print(type(e))
-        conn.commit()
+            conn.commit()
+        except (psycopg2.errors.UniqueViolation):
+            conn.commit()
+            return "repeat"
         
 
     def delete_from_employees_table(employee_id):
+        if employee_id not in [i[0] for i in employeesTable.show_employees_table()]:
+            conn.commit()
+            return "Not in table"
         cursor.execute("DELETE FROM employees WHERE id_employee=%s", (employee_id,))
         conn.commit()
 
     def update_employees_table(employee_id,surname, phone_number, edu_id):
+        if employee_id not in [i[0] for i in employeesTable.show_employees_table()]:
+            conn.commit()
+            return "Not in table"
         if surname:
             cursor.execute("UPDATE employees SET surname=%s WHERE id_employee=%s", (surname,employee_id))
         if phone_number:
@@ -130,16 +138,34 @@ class employeesTable():
         conn.commit()
         
 class stringsTable():
-    def show_string_a_table():
-        cursor.execute("SELECT * FROM string_a")
+    def show_string_a_table_with_ids():
+        cursor.execute("SELECT id_string, surname, assign_date, post, unit, salary FROM string_a JOIN employees ON string_a.id_employee = employees.id_employee JOIN positions ON string_a.id_post = positions.id_post \
+                        JOIN units ON string_a.id_units = units.id_units")
         lst = cursor.fetchall()
-        return sorted(lst)
+        return lst
+    
+    def show_string_a_table():
+        cursor.execute("SELECT surname, assign_date, post, unit, salary FROM string_a JOIN employees ON string_a.id_employee = employees.id_employee JOIN positions ON string_a.id_post = positions.id_post \
+                        JOIN units ON string_a.id_units = units.id_units")
+        lst = cursor.fetchall()
+        return lst
 
     def add_to_string_a_table(empl_id, assign_date, post_id, unit_id, salary):
-        cursor.execute("INSERT INTO string_a (id_employee, assign_date, id_post, id_units, salary) VALUES (%s,%s,%s, %s, %s)", (empl_id, assign_date, post_id, unit_id, salary))
-        conn.commit()
+        if not(assign_date) or not(datetime.strptime(assign_date, "%Y-%m-%d")) or datetime.strptime(assign_date, "%Y-%m-%d").date() > datetime.now().date():
+            return "Invalid date"
+        if salary == "":
+            return "Invalid salary"
+        try:
+            cursor.execute("INSERT INTO string_a (id_employee, assign_date, id_post, id_units, salary) VALUES (%s,%s,%s, %s,%s)", (empl_id, assign_date, post_id, unit_id, salary))
+            conn.commit()
+        except (psycopg2.errors.InvalidDatetimeFormat):
+            conn.commit()
+            return "Invalid date"
 
     def delete_from_string_a_table(string_id):
+        if string_id not in [i[0] for i in stringsTable.show_string_a_table_with_ids()]:
+            conn.commit()
+            return "Not in table"
         cursor.execute("DELETE FROM string_a WHERE id_string=%s", (string_id,))
         conn.commit()
